Add the build option to make the binaries debuggable
By default, the debug info will be stripped for the binaries, `cloudcore`
for example, to keep the binary as small as possible.

This behavior can be overwritten by adding the options `GOLDFLAGS` or
`GOGCFLAGS` when building the binaries.

As a example, the binary `cloudcore` built with the below command will
have DWARF and symbol table kept, the optimizations and inline will also
been disabled, this is helpful when you want to use the debugging tools
like `delve`.

`make all WHAT=cloudcore GOLDFLAGS="" GOGCFLAGS="-N -l"`

Signed-off-by: Dave Chen <dave.chen@arm.com>

Former-commit-id: b8f122e0bcd415028723a19123900e6a5b1417ce